requirements.txt
# Web & Core
fastapi
uvicorn>=0.20.0
pydantic_settings

# Database
sqlalchemy
psycopg[binary]

# Image & PDF Processing
pytesseract>=0.3.8
Pillow>=8.3.2
opencv-python-headless>=4.5.5.64
pymupdf>=1.19.0
libmagic
opencv-python-headless>=4.5.5.64
docling

# Data Science & Math
numpy>=1.20.0
scipy>=1.7.1
pandas>=1.3.5
matplotlib>=3.5.1
seaborn>=0.11.2
transformers
pgvector
sentence_transformers
pytesseract
python-multipart
torch
accelerate
python-magic
python-dateutil
pytz
openai
llama_cloud
lingua-language-detector 
langdetect
deepl